backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.57.0
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-redis==6.0.0
django-tenants==3.5.0
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==10.4.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.27.0
setuptools==80.9.0
shortuuid==1.0.11
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.6.0
python-barcode>=0.14.0
qrcode>=7.4.0   
django-storages>=1.13.0 
stripe>=5.5.0               
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5
redoc==0.2.1
django-filter>=23.0
